refactor(forms): move files in preparation for rename
Renames several files related to the "control" directive to reflect its
new name, the "field" directive
