+++ packages/forms/signals/src/api/field_directive.ts
@@ -10,9 +10,9 @@ import {
   computed,
   Directive,
   effect,
+  inject,
   InjectionToken,
   Injector,
-  inject,
   input,
   ɵCONTROL,
   ɵControl,
